Rename _get_component to _get_sdk_component
